Moved everything to MainActivity. Now I need to correct some things...
